Moving chatmanager to its own file.
